Move TmpTS sprite loading into Mods.TS.
