+++ OpenRA.Mods.TS/OpenRA.Mods.TS.csproj
@@ -56,6 +56,7 @@
     <Compile Include="TiberianSunRefinery.cs" />
     <Compile Include="Render\WithVoxelWalkerBody.cs" />
     <Compile Include="Render\WithVoxelUnloadBody.cs" />
+    <Compile Include="SpriteLoaders\TmpTSLoader.cs" />
   </ItemGroup>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
